+++ src/Response/YouTrackResponse.php
@@ -41,24 +41,24 @@ class YouTrackResponse implements ResponseContract
     }
 
     /**
-     * Get original HTTP client response.
+     * Returns original HTTP client response.
      *
      * @return \Psr\Http\Message\ResponseInterface
      */
-    public function getResponse(): ResponseInterface
+    public function httpResponse(): ResponseInterface
     {
         return $this->response;
     }
 
     /**
-     * Gets the response status code.
+     * Returns the response status code.
      *
      * The status code is a 3-digit integer result code of the server's attempt
      * to understand and satisfy the request.
      *
      * @return int
      */
-    public function getStatusCode(): int
+    public function statusCode(): int
     {
         return $this->response->getStatusCode();
     }
@@ -68,17 +68,17 @@ class YouTrackResponse implements ResponseContract
      *
      * @return string
      */
-    public function getCookie(): string
+    public function cookie(): string
     {
         return $this->response->getHeaderLine('Set-Cookie');
     }
 
     /**
-     * Gets the response Location header.
+     * Returns the response Location header.
      *
      * @return string
      */
-    public function getLocation(): string
+    public function location(): string
     {
         return $this->response->getHeaderLine('Location');
     }